+++ pandas/tests/io/formats/style/test_exceptions.py
@@ -3,6 +3,7 @@ import pytest
 jinja2 = pytest.importorskip("jinja2")
 
 from pandas import DataFrame
+import pandas._testing as tm
 
 from pandas.io.formats.style import Styler
 
@@ -31,3 +32,25 @@ def test_footer_bad_length(styler, kwarg, expected):
     msg = f"{expected} must have same length as ``func``"
     with pytest.raises(ValueError, match=msg):
         styler.set_footer(func=["mean"], **kwarg)
+
+
+def test_set_footer_warn():
+    df = DataFrame([["a"]])
+    styler = df.style.set_footer(["mean"], errors="warn")
+    msg = (
+        "`Styler.set_footer` raised Exception when calculating method `mean` on "
+        "column `0`"
+    )
+    with tm.assert_produces_warning(Warning, match=msg):
+        styler._translate(True, True)
+
+
+def test_set_footer_raise():
+    df = DataFrame([["a"]])
+    styler = df.style.set_footer(["mean"], errors="raise")
+    msg = (
+        "`Styler.set_footer` raised Exception when calculating method `mean` on "
+        "column `0`"
+    )
+    with pytest.raises(Exception, match=msg):
+        styler._translate(True, True)
